llm_tikg_integration/model/evaluation.py
"""
Model evaluation module for LLM-TIKG integration.
Handles model evaluation and metrics calculation.
"""

from typing import Dict, List, Any
import torch
from transformers import PreTrainedModel
import numpy as np
from sklearn.metrics import precision_recall_fscore_support
import logging

logger = logging.getLogger(__name__)

class ModelEvaluator:
    def __init__(self, model: PreTrainedModel, config: Dict[str, Any]):
        """
        Initialize model evaluator.
        
        Args:
            model (PreTrainedModel): Model to evaluate
            config (Dict[str, Any]): Evaluation configuration
        """
        self.model = model
        self.config = config
        
    def evaluate(self, eval_dataset) -> Dict[str, float]:
        """
        Evaluate model on dataset.
        
        Args:
            eval_dataset: Evaluation dataset
            
        Returns:
            Dict[str, float]: Evaluation metrics
        """
        self.model.eval()
        all_predictions = []
        all_labels = []
        
        with torch.no_grad():
            for batch in eval_dataset:
                outputs = self.model(**batch)
                predictions = torch.argmax(outputs.logits, dim=-1)
                
                all_predictions.extend(predictions.cpu().numpy())
                all_labels.extend(batch["labels"].cpu().numpy())
                
        metrics = self._calculate_metrics(all_predictions, all_labels)
        return metrics
        
    def _calculate_metrics(self, predictions: List[int], labels: List[int]) -> Dict[str, float]:
        """
        Calculate evaluation metrics.
        
        Args:
            predictions (List[int]): Model predictions
            labels (List[int]): Ground truth labels
            
        Returns:
            Dict[str, float]: Calculated metrics
        """
        precision, recall, f1, _ = precision_recall_fscore_support(
            labels, predictions, average='weighted'
        )
        
        return {
            "precision": precision,
            "recall": recall,
            "f1": f1
        }
        
    def evaluate_entity_extraction(self, predictions: List[str], ground_truth: List[str]) -> Dict[str, float]:
        """
        Evaluate entity extraction performance.
        
        Args:
            predictions (List[str]): Predicted entities
            ground_truth (List[str]): Ground truth entities
            
        Returns:
            Dict[str, float]: Entity extraction metrics
        """
        # TODO: Implement entity extraction evaluation
        return {}
        
    def evaluate_relationship_extraction(self, predictions: List[str], ground_truth: List[str]) -> Dict[str, float]:
        """
        Evaluate relationship extraction performance.
        
        Args:
            predictions (List[str]): Predicted relationships
            ground_truth (List[str]): Ground truth relationships
            
        Returns:
            Dict[str, float]: Relationship extraction metrics
        """
        # TODO: Implement relationship extraction evaluation
        return {} 